refactor: extract shared app logic from iOS and Android platforms
Extract three pieces of duplicated logic into shared functions so each
platform only contains its own platform-specific code:

- lib.rs: install_panic_hook() — identical panic hook previously inline
  in both ios/gpui_app.rs and android/jni.rs

- app.rs: open_zedra_window() — the cfg!(feature = "preview") branch
  that opened either PreviewApp or ZedraApp, duplicated in both
  ios/gpui_app.rs and android/app.rs

- app.rs: process_qr_result() — decode QR endpoint addr, call
  set_pending_qr_addr, and signal_terminal_data. iOS was also missing
  the signal_terminal_data call; this fixes that.
