+++ crates/zedra/src/android/jni.rs
@@ -68,21 +68,7 @@ fn init_logging() {
                 .with_tag("zedra"),
         );
 
-        std::panic::set_hook(Box::new(|info| {
-            let payload = info
-                .payload()
-                .downcast_ref::<&str>()
-                .map(|s| s.to_string())
-                .or_else(|| info.payload().downcast_ref::<String>().cloned())
-                .unwrap_or_else(|| "Unknown panic".to_string());
-
-            let location = info
-                .location()
-                .map(|l| format!("{}:{}:{}", l.file(), l.line(), l.column()))
-                .unwrap_or_else(|| "unknown".to_string());
-
-            log::error!("PANIC at {}: {}", location, payload);
-        }));
+        crate::install_panic_hook();
     });
 }
